+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 
 ### Dependency updates
 
+## [12.2.0] - 2022-02-25
+
+### Added
+
+- `EmailSelector`: Added new component. ([@lorgan3](https://github.com/lorgan3) in [#2002](https://github.com/teamleadercrm/ui/pull/2002))
+- `MarketingButtonGroup`: Added new component. ([@jelledc](https://github.com/jelledc) in [#2003](https://github.com/teamleadercrm/ui/pull/2003))
+
 ## [12.1.3] - 2022-02-18
 
 ### Changed